feat: Add optional wrapping of translation with html tag
This feature adds an optional showTerms configuration option that defaults to false. When showTerms is true, t will return the translation wrapped in a span tag with the title attribute set to the text being translated. For example:
t(text) will return <span title="#{text}">#{translation}</span>
If you would like to disable that setting for a translation, use the tns (translate no show) function.
There is also a ts function to wrap the translation in a div.
Use cases:

tns is useful inside an attribute, for example an alt attribute of an img tag.
ts is useful when chaining with markdownify. For example:
{{ 'some-term' | tns | markdownify | ts }}
this would turn off the wrapping of the translation so that you are not sending html to markdownify, then wrap the result of markdownify in a div so you can retain the benefit of showTerms